major refactoring

+++ src/main/scala/fr/acinq/fc/app/db/HostedUpdatesDb.scala
@@ -37,8 +37,8 @@ class HostedUpdatesDb(val db: PostgresProfile.backend.Database) {
     if (update.channelFlags.isNode1) addUpdate1(update) else addUpdate2(update)
 
   private def addUpdate1(update: ChannelUpdate): SqlAction[Int, PostgresProfile.api.NoStream, Effect] =
-    Updates.update1st(update.shortChannelId.toLong, channelUpdateCodec.encode(update).require.toHex, update.timestamp)
+    Updates.update1st(update.shortChannelId.toLong, channelUpdateCodec.encode(update).require.toHex, update.timestamp.toLong)
 
   private def addUpdate2(update: ChannelUpdate): SqlAction[Int, PostgresProfile.api.NoStream, Effect] =
-    Updates.update2nd(update.shortChannelId.toLong, channelUpdateCodec.encode(update).require.toHex, update.timestamp)
+    Updates.update2nd(update.shortChannelId.toLong, channelUpdateCodec.encode(update).require.toHex, update.timestamp.toLong)
 }
